+++ test/core_test_app/app/components/component_loader/javascripts/component_loader.js
@@ -66,5 +66,9 @@
     });
     w.show();
     this.netzkeLoadComponent({name: 'component_loaded_in_window', container: w});
+  },
+
+  onInaccessible: function() {
+    this.netzkeLoadComponent({name: 'inaccessible', container: this});
   }
 }